add new logo, fix types , delete size in container xxl, add mb for content in Hero
