+++ src/components/ui/HeroTitle/types.ts
@@ -1,0 +1,4 @@
+export interface HeroTitleType {
+  title: string;
+  className?: string;
+}
