+++ src/components/ui/Logo/types.ts
@@ -1,0 +1,4 @@
+export interface LogoType {
+  className?: string;
+  onClick?: () => void;
+}
